你好

+++ custom_components/lipro/core/api/request_policy.py
@@ -233,8 +233,8 @@ def compute_exponential_retry_wait_time(
     min_delay_seconds: float = 0.1,
 ) -> float:
     """Compute one exponential retry delay with optional min/max caps."""
-    min_delay = max(0.0, min_delay_seconds)
-    wait_time = max(min_delay, base_delay_seconds * (2**retry_count))
+    min_delay = float(max(0.0, min_delay_seconds))
+    wait_time = float(max(min_delay, base_delay_seconds * (2**retry_count)))
     if max_delay_seconds is None:
         return wait_time
-    return min(max_delay_seconds, wait_time)
+    return float(min(max_delay_seconds, wait_time))
